organized the sftp services
