+++ DotNet/DataAcquisition.Domain/Application/Services/Sftp/Processors/ISftpAcquisitionProcessor.cs
@@ -3,7 +3,7 @@ using LantanaGroup.Link.DataAcquisition.Domain.Application.Models.Domain;
 using LantanaGroup.Link.DataAcquisition.Domain.Infrastructure.Entities;
 using LantanaGroup.Link.DataAcquisition.Domain.Infrastructure.Models.Enums;
 
-namespace LantanaGroup.Link.DataAcquisition.Domain.Application.Services.Sftp;
+namespace LantanaGroup.Link.DataAcquisition.Domain.Application.Services.Sftp.Processors;
 
 /// <summary>
 /// Defines operations for processing SFTP acquisition requests.
